Rename: analytic_unit_creator -> utils
